requirements.txt
# AlphaFlow US — Python Dependencies
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
asyncpg>=0.29.0,<1.0.0
aiohttp>=3.9.0,<4.0.0
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0
yfinance>=0.2.30,<1.0.0
mplfinance>=0.12.10,<1.0.0
alpaca-py>=0.21.0,<1.0.0
finnhub-python>=2.4.0,<3.0.0
feedparser>=6.0.0,<7.0.0
beautifulsoup4>=4.12.0,<5.0.0
chromadb>=0.4.0,<1.0.0
langchain>=0.1.0,<1.0.0
langchain-ollama>=0.1.0,<1.0.0
langchain-chroma>=0.1.0,<1.0.0
ollama>=0.3.0,<1.0.0
apscheduler>=3.10.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.0,<1.0.0
Pillow>=10.1.0,<11.0.0
pydantic-settings>=2.1.0,<3.0.0
